move the create wallet action into common step
